task(#446): Cleanup compiler warnings
- Mostly from clang, but a few from icpc
- Mostly: explicit constructor warnings, implicit double -> int casting.
- Added #pragmas to silence compiler warnings for argos hooks, which I have no
  control over.

+++ include/fordyca/params/depth0/exec_estimates_params.hpp
@@ -46,7 +46,7 @@ NS_START(fordyca, params, depth0);
 struct exec_estimates_params : public rcppsw::params::base_params {
   bool enabled{false};
 
-  argos::CRange<double> generalist_range{};
+  argos::CRange<uint> generalist_range{};
 };
 
 NS_END(depth0, params, fordyca);
